+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Reduced support for BitBox01
 - Fix a bug that would prevent the app to perform firmware upgrade when offline.
 - Replace sidebar with bottom navigation bar for mobile devices
+- Added support to show on the BitBox when a transaction's recipient is an address of a different account on the device.
 
 # v4.47.2
 - Linux: fix compatiblity with some versions of Mesa that are incompatible with the bundled wayland libraries